+++ src/EditorFeatures/Core/EmbeddedLanguages/AbstractEmbeddedLanguageEditorFeaturesProvider.cs
@@ -3,18 +3,21 @@
 using System.Collections.Immutable;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.VirtualChars;
+using Microsoft.CodeAnalysis.Features.EmbeddedLanguages;
+using Microsoft.CodeAnalysis.Features.EmbeddedLanguages.RegularExpressions;
 using Microsoft.CodeAnalysis.LanguageServices;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
+namespace Microsoft.CodeAnalysis.Editor.EmbeddedLanguages
 {
     /// <summary>
     /// Abstract implementation of the C# and VB embedded language providers.
     /// </summary>
-    internal abstract class AbstractEditorFeaturesEmbeddedLanguagesProvider : AbstractFeaturesEmbeddedLanguagesProvider, IEditorFeaturesEmbeddedLanguagesProvider
+    internal abstract class AbstractEmbeddedLanguageEditorFeaturesProvider 
+        : AbstractEmbeddedLanguageFeaturesProvider, IEmbeddedLanguageEditorFeaturesProvider
     {
-        private readonly ImmutableArray<IEditorFeaturesEmbeddedLanguage> _embeddedLanguages;
+        private readonly ImmutableArray<IEmbeddedLanguageEditorFeatures> _embeddedLanguages;
          
-        protected AbstractEditorFeaturesEmbeddedLanguagesProvider(
+        protected AbstractEmbeddedLanguageEditorFeaturesProvider(
             int stringLiteralTokenKind,
             int interpolatedTextTokenKind,
             ISyntaxFactsService syntaxFacts,
@@ -22,11 +25,11 @@ namespace Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices
             IVirtualCharService virtualCharService)
             : base(stringLiteralTokenKind, interpolatedTextTokenKind, syntaxFacts, semanticFacts, virtualCharService)
         {
-            _embeddedLanguages = ImmutableArray.Create<IEditorFeaturesEmbeddedLanguage>(
-                new RegexEditorFeaturesEmbeddedLanguage(stringLiteralTokenKind, syntaxFacts, semanticFacts, virtualCharService));
+            _embeddedLanguages = ImmutableArray.Create<IEmbeddedLanguageEditorFeatures>(
+                new RegexEmbeddedLanguageEditorFeatures(stringLiteralTokenKind, syntaxFacts, semanticFacts, virtualCharService));
         }
 
-        public new ImmutableArray<IEditorFeaturesEmbeddedLanguage> GetEmbeddedLanguages()
+        public new ImmutableArray<IEmbeddedLanguageEditorFeatures> GetEmbeddedLanguages()
             => _embeddedLanguages;
     }
 }
