+++ src/EditorFeatures/Core/EmbeddedLanguages/RegularExpressions/RegexEmbeddedBraceMatcher.cs
@@ -6,12 +6,14 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.Common;
+using Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions;
+using Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.VirtualChars;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.Shared.Extensions;
 using Roslyn.Utilities;
 
-namespace Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices
+namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.RegularExpressions
 {
     using RegexToken = EmbeddedSyntaxToken<RegexKind>;
     using RegexTrivia = EmbeddedSyntaxTrivia<RegexKind>;
@@ -21,9 +23,9 @@ namespace Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageSe
     /// </summary>
     internal sealed class RegexEmbeddedBraceMatcher : IBraceMatcher
     {
-        private readonly RegexEmbeddedLanguage _language;
+        private readonly RegexEmbeddedLanguageEditorFeatures _language;
 
-        public RegexEmbeddedBraceMatcher(RegexEmbeddedLanguage language)
+        public RegexEmbeddedBraceMatcher(RegexEmbeddedLanguageEditorFeatures language)
         {
             _language = language;
         }
