+++ src/Features/Core/Portable/DocumentHighlighting/AbstractDocumentHighlightsService.cs
@@ -9,6 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
 using Microsoft.CodeAnalysis.ErrorReporting;
+using Microsoft.CodeAnalysis.Features.EmbeddedLanguages;
 using Microsoft.CodeAnalysis.FindSymbols;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.PooledObjects;
@@ -95,7 +96,7 @@ namespace Microsoft.CodeAnalysis.DocumentHighlighting
         private async Task<ImmutableArray<DocumentHighlights>> TryGetEmbeddedLanguageHighlightsAsync(
             Document document, int position, IImmutableSet<Document> documentsToSearch, CancellationToken cancellationToken)
         {
-            var languagesProvider = document.GetLanguageService<IFeaturesEmbeddedLanguagesProvider>();
+            var languagesProvider = document.GetLanguageService<IEmbeddedLanguageFeaturesProvider>();
             if (languagesProvider != null)
             {
                 foreach (var language in languagesProvider.GetEmbeddedLanguages())
